clean code. reusable address component
