+++ src/app/views/address/address.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../scss/variables';
+@import '../../../scss/variables';
 
 .main-h4 {
   @extend %roboto-regular;
