--- src/app/views/address/borrowed-loans/borrowed-loans.component.html
@@ -1,62 +1,0 @@
-<ngx-spinner
-  bdColor="rgba(0,0,0,.3)"
-  size="medium"
-  color="#fff"
-  type="ball-clip-rotate-multiple"
-></ngx-spinner>
-
-<div
-  class="body-scrollable activity-wrapper"
-  #scrollingBlock
->
-  <div class="container">
-    <div class="row align-items-center pt-2 pt-lg-0">
-
-      <div class="col-sm-12 col-lg-6 d-none d-lg-block">
-        <div class="title main-h1">borrowed by</div>
-        <div class="title main-h4 ellipsis my-2">
-          {{ address }}
-        </div>
-        <mat-chip-list>
-          <mat-chip
-            class="mat-chip--gray margintop30 marginright15"
-            selected="true"
-          >
-            {{ loans.length }} BORROWED
-          </mat-chip>
-        </mat-chip-list>
-      </div>
-
-      <div class="
-        col-sm-12 col-lg-6 d-flex justify-content-center justify-content-md-end
-      ">
-        <app-my-loans-tabs></app-my-loans-tabs>
-      </div>
-
-    </div>
-    <div class="loan-container row">
-
-      <div class="col-sm-12 text-center margintop30" *ngIf='!availableLoans'>
-        <div class="title main-h1">This address is not a borrower</div>
-      </div>
-
-      <virtual-scroller
-        #scroll
-        [items]="loans || []"
-        [parentScroll]="scrollingBlock"
-        [bufferAmount]="4"
-        [scrollThrottlingTime]="500"
-      >
-        <div
-          class="col-sm-12 col-md-6 col-lg-4 col-xl-3 card-col mb-5"
-          *ngFor="let loan of scroll.viewPortItems"
-        >
-          <app-loan-card [loan]="loan"></app-loan-card>
-        </div>
-      </virtual-scroller>
-
-    </div>
-  </div>
-
-  <app-footer></app-footer>
-</div>
