--- src/app/views/address/lent-loans/lent-loans.component.scss
@@ -1,11 +1,0 @@
-@import '../../../../scss/variables';
-
-.main-h4 {
-  @extend %roboto-regular;
-}
-
-virtual-scroller {
-  ::ng-deep .scrollable-content {
-    height: auto;
-  }
-}
